+++ frontend/src/main.tsx
@@ -3,6 +3,9 @@ import "./index.css";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { I18nextProvider } from "react-i18next";
+import { Provider } from "react-redux";
+
+import { store } from "store/store";
 
 import i18n from "i18n/index";
 
@@ -16,8 +19,10 @@ if (!rootElement) {
 
 createRoot(rootElement).render(
     <StrictMode>
-        <I18nextProvider i18n={i18n}>
-            <App />
-        </I18nextProvider>
+        <Provider store={store}>
+            <I18nextProvider i18n={i18n}>
+                <App />
+            </I18nextProvider>
+        </Provider>
     </StrictMode>,
 );